% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{areas_pi}
\alias{areas_pi}
\title{Catchment (roof) areas of the IFPB-PI campus}
\format{
A data frame with one row per building:
\describe{
\item{block}{Building name.}
\item{area_m2}{Projected roof area, in square metres (m2).}
}
}
\source{
Vectorised aerial imagery of the IFPB, Campus Princesa Isabel.
}
\usage{
areas_pi
}
\description{
Projected roof areas of the four buildings of the IFPB, Campus Princesa
Isabel, that can serve as rainwater catchment surfaces.
}
\keyword{datasets}
