% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{rh_demo}
\alias{rh_demo}
\title{Open the rharv demonstration notebook}
\usage{
rh_demo()
}
\value{
Invisibly, the path to the demo file.
}
\description{
Opens the bundled R Markdown demo (\code{rharv-demo.Rmd}), which exercises every
function, plot and dataset of the package, in your editor (RStudio when run
from an RStudio session). Use Knit to render it to HTML.
}
\examples{
\dontrun{
rh_demo()
}
}
