% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rharv-package.R
\docType{package}
\name{rharv-package}
\alias{rharv}
\alias{rharv-package}
\title{rharv: Daily Water-Balance Simulation of Rainwater Harvesting Systems}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simulates rainwater harvesting reservoirs with a daily water-balance (mass-balance) model. From a daily precipitation series it computes the captured volume, runs the reservoir balance and derives performance metrics and guarantee-based sizing values (constant demand, catchment area and reservoir capacity). The available-volume relation and the daily-simulation approach follow established references (e.g. Souza, 2015; ABNT NBR 15527). Aimed at water-security analysis of buildings, especially in semi-arid regions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://arturlourenco.github.io/rharv}
  \item \url{https://github.com/ArturLourenco/rharv}
  \item Report bugs at \url{https://github.com/ArturLourenco/rharv/issues}
}

}
\author{
\strong{Maintainer}: Artur Moises Goncalves Lourenco \email{artur.lourenco@ifpb.edu.br}

}
\keyword{internal}
